+++ src/chassis.cpp
@@ -7,12 +7,12 @@
 #define kI 0.5
 #define kD 0.5
 
-void chassis::drive(float distance) {
+void chassis::drive(double distance) {
 	PIDController drive_pid(kP, kI, kD);
 	PIDController turn_pid(kP, kI, kD);
 
-	float drive;
-	float turn;
+	double drive;
+	double turn;
 
 	while (drive_pid.get_error() != 0) {
 		drive = drive_pid.calculate(distance, odom::get_x());
@@ -26,10 +26,10 @@ void chassis::drive(float distance) {
 	drive_right.brake();
 }
 
-void chassis::turn_abs(float heading) {
+void chassis::turn_abs(double heading) {
 	PIDController pid(kP, kI, kD);
-	float output;
-	float dir = 1;
+	double output;
+	double dir = 1;
 	double current_hdg = imu.get_heading();
 
 	if ((heading - current_hdg) > 180) dir = -1;
@@ -45,8 +45,8 @@ void chassis::turn_abs(float heading) {
 	drive_right.brake();
 }
 
-void chassis::turn_rel(float degrees) {
-	float heading = imu.get_heading() + degrees;
+void chassis::turn_rel(double degrees) {
+	double heading = imu.get_heading() + degrees;
 	if (heading > 180) heading -= 360;
 	if (heading < 0) heading = fabs(heading);
